Move commands to pkg, so cmd only contains minimalist main.go
