+++ pkg/kn/commands/service_list.go
